+++ lib/ts/recipe/thirdparty/components/themes/styles.ts
@@ -73,14 +73,14 @@ export function getStyles(palette: NormalisedPalette): NormalisedDefaultStyles {
             justifyContent: "center",
         },
 
-        providerGoogle: getButtonStyle(providerColors.google),
-        providerGithub: getButtonStyle(providerColors.github, true),
-        providerTwitter: getButtonStyle(providerColors.twitter),
-        providerFacebook: getButtonStyle(providerColors.facebook),
-        providerApple: getButtonStyle(providerColors.apple, true),
+        providerGoogle: getButtonStyle(providerColors.google, palette.colors.buttonText),
+        providerGithub: getButtonStyle(providerColors.github, palette.colors.buttonText, true),
+        providerTwitter: getButtonStyle(providerColors.twitter, palette.colors.buttonText),
+        providerFacebook: getButtonStyle(providerColors.facebook, palette.colors.buttonText),
+        providerApple: getButtonStyle(providerColors.apple, palette.colors.buttonText, true),
 
         providerCustom: {
-            ...getButtonStyle(providerColors.custom),
+            ...getButtonStyle(providerColors.custom, palette.colors.buttonText),
             color: "#000",
             border: "1px solid #000",
             "&:active": {